Revert parallel_for integration in low level transformation
In dcbe5a507cd6eedce08f27087d6e3d33f1310485, we tried to address a
memory leak issue when using intra_threads = 1 (possibly related to
nested parallel regions). However, the fix is incorrect as
omp_get_num_threads returns a different value in the parallel region.

As a quick and safe fix, we revert the parallel_for integration for
now.