Fix parallel translations from multiple Python threads
An exclusive lock was added to make the methods `translate` and
`{load,unload}_model` mutually exclusive. However, this also makes
concurrent calls to `translate` mutually exclusive.

The true fix would be to use a shared lock but this proved to be a bit
complex to add without C++17 support.

For simplicity, we remove this locking mechanism for now and defer
this responsibility to the application. This is also consistent with
`del translator` which was never thread safe.