Upgrade to GCC 8 in CentOS 7 images to build the cli
We removed the cli in be25c57b696a67d7398190ed5b723808a805c266 because
compiling the library with devtoolset-3 seemed to cause issues in a
downstream Docker image. This is apparently no longer an issue when
using a more recent devtoolset.

Note that nvcc still requires the GCC version that is shipped with
CentOS 7 (see `CUDA_HOST_COMPILER`).