On GPU, always expand to beam size before the first decoding step
This is a partial revert of 7ebba2e8 which seems to cause a
performance regression when using large beam size on GPU.

The performance regression can be observed on both FP32 and FP16
decoding, but not with the same beam size. This could indicate that
the issue is related to the caching allocator that does not like the
larger batch size at the second step.

The behavior is unchanged on CPU.