Penalize EOS symbol in secondary beams after the target prefix
At the first unconstrained decoding step, only the first beam is
expanded. It happens that the EOS symbol appears in the topk,
especially when k is large. This can produce incorrect and short
predictions that dominate others (when `length_penalty` is not used).
To mitigate this issue, we penalize the EOS token but only in
secondary beams.