Fix and disable padding to a multiple of 16 for INT8
The previous logic was incorrect and the padding was not applied. The
code checked the layers output type which can only be float32 or
float16. Instead, it should check the global compute type.

After fixing this issue, it appears padding to a multiple of 16 does
not help. So let's disable it for now and gather more data.